+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Align header with Shopify
 - Add shop link to header
+- Align footer with Shopify
 
 ---